(GH-81) Added VS-HelpLink and Links in message
since the update of Microsoft.Build.Utilities.Core to v16.8.0
it is possible to add a HelpLink to errors and warnings
so in Visual Studio (beginning with 16.8) clickable links
will be rendered in the error pane.

Additionally for non VS users (or users of older VS)
the link is added to the output of the message.